fix: move component to presentation
